requirements.txt
black==23.3.0
blinker==1.6.2
click==8.1.3
colorama==0.4.6
cssbeautifier==1.14.7
djlint==1.25.0
EditorConfig==0.12.3
Flask==2.3.1
html-tag-names==0.1.2
html-void-elements==0.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsbeautifier==1.14.7
MarkupSafe==2.1.2
mypy-extensions==1.0.0
mysql-connector-python==8.0.33
nodeenv==1.7.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.3.0
protobuf==3.20.3
pyright==1.1.305
python-dotenv==1.0.0
PyYAML==6.0
regex==2023.3.23
six==1.16.0
tqdm==4.65.0
Werkzeug==2.3.0
